third_party/WebKit/LayoutTests/inspector-protocol/css/pseudo-element-matching-selectors.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script>

function addBeforeElement()
{
    document.getElementById("style").textContent = "#for-pseudo:before { content: \"BEFORE\" }";
}

function test()
{
    var nodeInfo = {};
    var childrenCallback;

    InspectorTest.eventHandler["DOM.setChildNodes"] = setChildNodes;
    getDocument();

    function getDocument()
    {
        step({
            name: "Get the Document",
            command: "DOM.getDocument",
            parameters: {},
            callback: getImmediateChildren
        });
    };

    function getImmediateChildren(result)
    {
        var bodyId = result.root.children[0].children[1].nodeId;
        childrenCallback = onChildrenRequested;
        step({
            name: "Get immediate children of the body",
            command: "DOM.requestChildNodes",
            parameters: {"nodeId": bodyId}
        });
    };

    function onChildrenRequested()
    {
        var node = findNodeById("for-pseudo");
        var beforeNode = node.pseudoElements[0];
        step({
            name: "Request matching styles for #for-pseudo::before",
            command: "CSS.getMatchedStylesForNode",
            parameters: {nodeId: beforeNode.nodeId},
            callback: stylesReceived
        });
    }

    function stylesReceived(result)
    {
        var matchedRules = result.matchedCSSRules;
        for (var i = 0; i < matchedRules.length; ++i) {
            var match = matchedRules[i];
            if (match.rule.selectorList.text === "#for-pseudo::before") {
                InspectorTest.log("#for-pseudo::before matching the :before element: " + (match.matchingSelectors[0] === 0));
                InspectorTest.completeTest();
                return;
            }
        }
        InspectorTest.log("#for-pseudo::before rule not received");
        InspectorTest.completeTest();
    }

    function setChildNodes(message)
    {
        var nodes = message.params.nodes;
        for (var i = 0; i < nodes.length; ++i)
            addNodesRecursive(nodes[i]);
        var callback = childrenCallback;
        childrenCallback = null;
        if (callback)
            callback();
    }

    function step(test)
    {
        InspectorTest.log("\n=== " + test.name + " ===\n");
        InspectorTest.sendCommand(test.command, test.parameters, function(messageObject) {
            if (messageObject.hasOwnProperty("error"))
                InspectorTest.log("Backend error: " + messageObject.error.message + " (" + messageObject.error.code + ")\n");

            if (test.callback)
                test.callback(messageObject.result);
        });
    }

    function findNodeById(id)
    {
        for (var nodeId in nodeInfo) {
            var node = nodeInfo[nodeId];
            var attrs = node.attributes;
            if (!attrs)
                continue;
            for (var i = 0; i < attrs.length; i += 2) {
                var name = attrs[i];
                if (name !== "id")
                    continue;
                if (attrs[i + 1]  === id)
                    return node;
            }
        }
        return null;
    }

    function addNodesRecursive(root)
    {
        addNode(root);
        if (!root.children)
            return;
        for (var i = 0; i < root.children.length; ++i)
            addNodesRecursive(root.children[i]);
    }

    function addNode(node)
    {
        nodeInfo[node.nodeId] = node;
    }
}

</script>
<style>
#for-pseudo:before {
  color: red;
  content: "BEFORE";
}
</style>
</head>
<body onload="runTest()">
<div id="for-pseudo">Test</div>
</body>
</html>
