third_party/WebKit/LayoutTests/inspector-protocol/css/pseudo-element-matching-selectors-expected.txt

Test

=== Get the Document ===


=== Get immediate children of the body ===


=== Request matching styles for #for-pseudo::before ===

#for-pseudo::before matching the :before element: true
